reworked basically everything. Much neater now

+++ src/main/Java/Randomiser/RandomiserFactory.java
@@ -6,5 +6,5 @@ package Java.Randomiser;
 public interface RandomiserFactory {
     Randomiser createRandomiser();
 
-    void update(String randomiserType, int[] buildData);
+    void update(String randomiserType, long[] buildData);
 }
